Cleaned up main file, moved edited send func to over.py
