+++ bot/utils/over.py
@@ -5,7 +5,9 @@ import inspect
 import io
 import itertools
 
+import discord
 from discord.ext.commands import HelpFormatter as HelpF, Paginator, Command
+from ghostbin import GhostBin
 
 from bot.utils.args import ArgParseConverter as ArgPC
 
@@ -98,3 +100,18 @@ class HelpFormatter(HelpF):
         ending_note = self.get_ending_note()
         self._paginator.add_line(ending_note)
         return self._paginator.pages
+
+
+old_send = discord.abc.Messageable.send
+
+
+async def send(self, content=None, **kwargs):
+    """Overrides default send method in order to create a paste if the response is more than 2000 characters"""
+    if content is not None and any(x in str(content) for x in ["@everyone", "@here"]):
+        content = content.replace("@everyone", "@\u0435veryone").replace("@here", "@h\u0435re")
+    if content is not None and len(str(content)) > 2000:
+        paste = await GhostBin().paste(content, expire="15m")
+        await old_send(self, f"Hey, I couldn't handle all the text I was gonna send you, so I put it in a paste!"
+                             f"\nThe link is **{paste}**, but it expires in 15 minutes, so get it quick!", **kwargs)
+    else:
+        await old_send(self, content, **kwargs)
